Exponential Search in Java: Algorithm, Implementation & Analysis # **Exponential Search**
An **exponential search** is a combination of two methods:

- Find the range in which the element exists.
- Perform a **binary search** in that range.

To begin a search, we find the range. We do this by first checking to see if the desired element is in the first position. If not, we try an array size of 2, then 4, then 6... and so on. If the last element in the partial array is not greater than the element, we perform a binary search.

Think of a binary search as a chopping of the array. First, we look at the whole array, then chop it in half. If the value is less than the middle, we look at the lower half. If it is greater than the middle, we look at the upper half. This is repeated until we find the value or the list is empty.

So what would this look like in Java code? Let's take a look and break down each step.

Note: In order for this to work the array must be sorted!